Cellula_4week_Omar_Arnous/requirements.txt
dotenv
huggingface_hub
langchain_chroma
langchain_openai
langchain_core
python-telegram-bot
nltk
wikipedia
numpy
deepeval